requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
astrapy==2.0.1 
python-dotenv==1.1.0 
pydantic==2.11.4 
numpy==2.2.5 
langgraph==0.4.1 
langchain-core==0.3.56
langgraph-checkpoint==2.0.25